requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.9
torch==2.10.0
numpy==2.2.6
pillow==12.1.0
pandas==2.1.4
scikit-learn==1.3.2
seaborn==0.13.0
matplotlib==3.8.2
requests==2.31.0
joblib==1.3.2
mlflow==2.10.0
prometheus-client==0.19.0
pytest==7.4.3
